Split parameter classes into separate files
* Sorted headers and sources in CMakeLists.txt
