+++ src/complex/Filter/ConstDataParameter.cpp
@@ -1,0 +1,9 @@
+#include "ConstDataParameter.hpp"
+
+namespace complex
+{
+ConstDataParameter::Mutability ConstDataParameter::mutability() const
+{
+  return Mutability::Const;
+}
+} // namespace complex
