+++ src/complex/Filter/IParameter.hpp
@@ -12,9 +12,6 @@
 #include "complex/Filter/Arguments.hpp"
 #include "complex/complex_export.hpp"
 
-#include "complex/DataStructure/DataPath.hpp"
-#include "complex/DataStructure/DataStructure.hpp"
-
 namespace complex
 {
 /**
@@ -99,135 +96,5 @@ public:
 
 protected:
   IParameter() = default;
-};
-
-class COMPLEX_EXPORT AbstractParameter : public IParameter
-{
-public:
-  ~AbstractParameter() noexcept override = default;
-
-  AbstractParameter(const AbstractParameter& other) = delete;
-  AbstractParameter(AbstractParameter&& other) noexcept = delete;
-
-  AbstractParameter& operator=(const AbstractParameter& other) = delete;
-  AbstractParameter& operator=(AbstractParameter&&) noexcept = delete;
-
-  [[nodiscard]] std::string name() const final;
-
-  [[nodiscard]] std::string humanName() const final;
-
-  [[nodiscard]] std::string helpText() const final;
-
-  [[nodiscard]] std::optional<std::any> defaultValue() const final;
-
-protected:
-  AbstractParameter() = delete;
-
-  AbstractParameter(const std::string& name, const std::string& humanName, const std::string& helpText, const std::optional<std::any>& defaultValue = {});
-
-private:
-  std::string m_Name;
-  std::string m_HumanName;
-  std::string m_HelpText;
-  std::optional<std::any> m_DefaultValue;
-};
-
-class COMPLEX_EXPORT ValueParameter : public AbstractParameter
-{
-public:
-  ~ValueParameter() noexcept override = default;
-
-  ValueParameter(const ValueParameter&) = delete;
-  ValueParameter(ValueParameter&&) noexcept = delete;
-
-  ValueParameter& operator=(const ValueParameter&) = delete;
-  ValueParameter& operator=(ValueParameter&&) noexcept = delete;
-
-  [[nodiscard]] Type type() const final;
-
-  [[nodiscard]] virtual bool validate(const std::any& value) const = 0;
-
-protected:
-  ValueParameter() = delete;
-  using AbstractParameter::AbstractParameter;
-};
-
-class COMPLEX_EXPORT DataParameter : public AbstractParameter
-{
-public:
-  enum class Category : u8
-  {
-    Required = 0,
-    Created
-  };
-
-  enum class Mutability : u8
-  {
-    Const = 0,
-    Mutable
-  };
-
-  ~DataParameter() noexcept override = default;
-
-  DataParameter(const DataParameter&) = delete;
-  DataParameter(DataParameter&&) noexcept = delete;
-
-  DataParameter& operator=(const DataParameter&) = delete;
-  DataParameter& operator=(DataParameter&&) noexcept = delete;
-
-  [[nodiscard]] Type type() const final;
-
-  [[nodiscard]] Category category() const;
-
-  [[nodiscard]] virtual bool validate(const DataStructure& dataStructure, const std::any& value) const = 0;
-
-  [[nodiscard]] virtual Mutability mutability() const = 0;
-
-protected:
-  DataParameter() = delete;
-  DataParameter(const std::string& name, const std::string& humanName, const std::string& helpText, const std::optional<std::any>& defaultValue, Category category);
-
-private:
-  Category m_Category;
-};
-
-class COMPLEX_EXPORT ConstDataParameter : public DataParameter
-{
-public:
-  ~ConstDataParameter() noexcept override = default;
-
-  ConstDataParameter(const ConstDataParameter&) = delete;
-  ConstDataParameter(ConstDataParameter&&) noexcept = delete;
-
-  ConstDataParameter& operator=(const ConstDataParameter&) = delete;
-  ConstDataParameter& operator=(ConstDataParameter&&) noexcept = delete;
-
-  [[nodiscard]] Mutability mutability() const final;
-
-  [[nodiscard]] virtual std::any resolve(const DataStructure& dataStructure, const std::any& value) const = 0;
-
-protected:
-  ConstDataParameter() = delete;
-  using DataParameter::DataParameter;
-};
-
-class COMPLEX_EXPORT MutableDataParameter : public DataParameter
-{
-public:
-  ~MutableDataParameter() noexcept override = default;
-
-  MutableDataParameter(const MutableDataParameter&) = delete;
-  MutableDataParameter(MutableDataParameter&&) noexcept = delete;
-
-  MutableDataParameter& operator=(const MutableDataParameter&) = delete;
-  MutableDataParameter& operator=(MutableDataParameter&&) noexcept = delete;
-
-  [[nodiscard]] Mutability mutability() const final;
-
-  [[nodiscard]] virtual std::any resolve(DataStructure& dataStructure, const std::any& value) const = 0;
-
-protected:
-  MutableDataParameter() = delete;
-  using DataParameter::DataParameter;
 };
 } // namespace complex
