+++ src/complex/Filter/MutableDataParameter.cpp
@@ -1,0 +1,9 @@
+#include "MutableDataParameter.hpp"
+
+namespace complex
+{
+MutableDataParameter::Mutability MutableDataParameter::mutability() const
+{
+  return Mutability::Mutable;
+}
+} // namespace complex
