+++ src/complex/Filter/ValueParameter.cpp
@@ -1,0 +1,9 @@
+#include "ValueParameter.hpp"
+
+namespace complex
+{
+ValueParameter::Type ValueParameter::type() const
+{
+  return Type::Value;
+}
+} // namespace complex
